Restructured React components, incoherences handling token when login and logout
